杨浦区地摊经济监测与点位决策支持系统/requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
streamlit>=1.43
folium>=0.19
streamlit-folium>=0.24
pydeck>=0.9
python-dotenv>=1.0
requests>=2.32
joblib>=1.4
pytest>=8.3
matplotlib>=3.9
playwright>=1.52
